Separated auth commands from login to separate imports.
