+++ src/lib/commands/index.ts
@@ -1,4 +1,5 @@
 import "./common";
+import "./auth";
 import "./general";
 import "./dates";
 import "./login";
